renamed powered by to protected by
