+++ locales/en/common.json
@@ -68,7 +68,7 @@
   "admin_portal": "Admin Portal",
   "previous_step": "Previous Step",
   "next_step": "Next Step",
-  "boxyhq_powered_by": "Powered by Ory",
+  "ory_protected_by": "Protected by Ory Polis",
   "guides": "Guides",
   "learn_to_enable_auth_methods": "Please visit <docLink>Ory Polis documentation</docLink> to learn how to enable the Magic Link or Email/Password authentication methods.",
   "advanced_sp_saml_configuration": "Advanced Service Provider (SP) SAML Configuration",
